Consertos

+++ receitas.php
@@ -42,7 +42,7 @@
 
                     $sql= "SELECT id_planta FROM post WHERE id_post='$i'";
                     $resultado = pg_query($connect, $sql);
-                    $planta_mae = pg_fetch_array($resultado)[0];
+                    $planta_mae = (int) pg_fetch_array($resultado)[0];
